Rearranged the computer page files, and testing NextJS routing. The computer pages need some cleaning up, and then I will be ready to add more pages.
