--- src/app/about/MyBuild/page.js
@@ -1,16 +1,0 @@
-"use client";
-
-import './MyBuild.css';
-
-export default function MyBuild() {
-    return (
-		<div className='container'>
-            <img loading="lazy" decoding="async" src="/images/myBuild/20241128_105614.jpg" alt="meeting" aria-hidden="true" />
-            <img loading="lazy" decoding="async" src="/images/myBuild/20241128_114024.jpg" alt="meeting" aria-hidden="true" />
-            <img loading="lazy" decoding="async" src="/images/myBuild/20241128_130847.jpg" alt="meeting" aria-hidden="true" />
-            <img loading="lazy" decoding="async" src="/images/myBuild/20241214_172531.jpg" alt="meeting" aria-hidden="true" />
-            <img loading="lazy" decoding="async" src="/images/myBuild/20241215_142804.jpg" alt="meeting" aria-hidden="true" />
-            <img loading="lazy" decoding="async" src="/images/myBuild/20241216_180049.jpg" alt="meeting" aria-hidden="true" />
-        </div>
-  	);
-}
