+++ src/app/about/NewComputer/MyBuild/MyBuild.css
@@ -5,7 +5,7 @@
     margin-top: 5rem;
 }
 
-.container > img{
+.container img{
     padding: 1rem;
     max-width: 800px;
 }
