+++ src/app/about/NewComputer/MyBuild/page.js
@@ -1,0 +1,29 @@
+"use client";
+
+import Link from "next/link";
+import './MyBuild.css';
+
+export default function MyBuild() {
+    return (
+		<div className='container'>
+			<Link href="/about/NewComputer">
+				<img loading="lazy" decoding="async" src="/images/myBuild/20241128_105614.jpg" alt="meeting" aria-hidden="true" />
+			</Link>
+			<Link href="/about/NewComputer">
+				<img loading="lazy" decoding="async" src="/images/myBuild/20241128_114024.jpg" alt="meeting" aria-hidden="true" />
+			</Link>
+			<Link href="/about/NewComputer">
+				<img loading="lazy" decoding="async" src="/images/myBuild/20241128_130847.jpg" alt="meeting" aria-hidden="true" />
+			</Link>
+			<Link href="/about/NewComputer">
+				<img loading="lazy" decoding="async" src="/images/myBuild/20241214_172531.jpg" alt="meeting" aria-hidden="true" />
+			</Link>
+			<Link href="/about/NewComputer">
+				<img loading="lazy" decoding="async" src="/images/myBuild/20241215_142804.jpg" alt="meeting" aria-hidden="true" />
+			</Link>
+			<Link href="/about/NewComputer">
+				<img loading="lazy" decoding="async" src="/images/myBuild/20241216_180049.jpg" alt="meeting" aria-hidden="true" />
+			</Link>
+        </div>
+  	);
+}
